[App Search] Automated Curations: Split Curations view into "Overview" and "Settings" tabs (#112488)
