+++ x-pack/plugins/enterprise_search/public/applications/test_helpers/index.ts
@@ -16,6 +16,7 @@ export {
   getPageDescription,
   getPageHeaderActions,
   getPageHeaderChildren,
+  getPageHeaderTabs,
 } from './get_page_header';
 
 // Misc
